Adds multi-module parent pom to enable automated building of threadfix subprojects.
The dependencies that scan agent and main had were being managed using
Eclipse-specific .classpath files. This was a bad solution that was tied
to an IDE and made it very hard to build these projects on the command
line. This becomes much easier when a multi-module build is used,
although we may run into some problems with this solution too down the
road.