Moves ThreadFix Entities into their own module.
Also updates the ThreadFix REST endpoints and clients to be typed RestResponse objects.

For example, the controller method for application information used by plugins returns
an object of type RestResponse<Application.Info[]>, and the client code then knows that
it can create such an object from the JSON. 

This leads to a tiny type-safe client library using the native ThreadFix types.