DGTF-2243 address WhiteHat performance issues
The solution here was to not do as many Date comparisons and to not copy all of a Finding's fields for each scan. Since WhiteHat findings contain the full XML of the finding, copying it (num scans) * (num findings) times leads to a lot of memory usage and GC cycles.