DGTF-2400 fix statistics for Scans
This commit is large because it includes some scans to use in unit tests.

Most of the new code is in HibernateScanDao.

This solves the problem we had before where vulnerabilities with findings merged
over different scanners had incorrect statistics because firstFindingForVuln was used in stats
calculation

This involves testing uniqueness over the following fields:
 - vuln ID
 - scan ID (time, so we can sort)
 - channel ID

Then sort by date and take the earliest one. It will then go through another SQL statement
to make sure it's appropriate. We also need to cache this for multiple runs.

Strategy-wise, we get all of these fields and hash them to create a key
This is a much smaller structure than hibernate objects. We want that to reduce the
memory footprint of this algorithm.

With a caveat: some findings have a higher priority than others. If the finding already has a
counter, we want to include it in the list so that another finding merged to that finding doesn't
also get a counter, leading to that vulnerability getting double counted. Also, if one is marked
"firstFindingForVuln" we probably want that one.