DGTF-2524 Fix add Defect Tracker for Defect Trackers without default credentials
The problem was that we decrypted the existing credentials for the application over the submitted ones in the normal validation method. It only overrides the new credentials because we changed the behavior to write the new fields onto the Hibernate-managed application to avoid overwriting other fields. That part works, but then in the validation method we load the application from Hibernate again. When we pass the second pointer to the decrypt method, it pointed to the same object as the first. It then writes to the same fields as the first.

We need to do this in a more idiomatic Spring way and split the EditApplicationController into 3 controllers. I have this in a branch now but we're about to release and I want to minimize changes.

For now I removed the normal validation (that decrypts the credentials) in the setDefectTracker method because it doesn't test any of the changed fields anyway.